Fix for broken behaviour in math.clamp
There is an issue in the implementation of math.clamp, which means the value to be clamped is returned rather than the lower bound, if a < value < b.

This commit fixes that behaviour, ensuring that the correct value is returned in the case of a < value < b.

Additionally, this commit defines the expected behaviour when a > b.

Tests included.